pyproject.toml
[project]
name = "ai-exam-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "google-generativeai>=0.8.6",
    "gtts>=2.5.4",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "sounddevice>=0.5.3",
    "speechrecognition>=3.14.5",
    "streamlit>=1.40.0",
]
